backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
websockets>=12.0
